+++ dejafu-tests/Cases/MultiThreaded.hs
@@ -14,7 +14,7 @@ import Test.HUnit.DejaFu (testDejafuWay)
 
 import Control.Concurrent.Classy
 import Control.Monad.STM.Class
-import Test.DejaFu.Conc (Conc, ConcST, subconcurrency)
+import Test.DejaFu.Conc (ConcT, ConcST, subconcurrency)
 
 import Utils
 
@@ -286,7 +286,7 @@ schedDaemon = do
 -- Subconcurrency
 
 -- | Subcomputation deadlocks sometimes.
-scDeadlock1 :: Monad n => Conc n r (Either Failure ())
+scDeadlock1 :: Monad n => ConcT r n (Either Failure ())
 scDeadlock1 = do
   var <- newEmptyMVar
   subconcurrency $ do
@@ -295,7 +295,7 @@ scDeadlock1 = do
 
 -- | Subcomputation deadlocks sometimes, and action after it still
 -- happens.
-scDeadlock2 :: Monad n => Conc n r (Either Failure (), ())
+scDeadlock2 :: Monad n => ConcT r n (Either Failure (), ())
 scDeadlock2 = do
   var <- newEmptyMVar
   res <- subconcurrency $ do
@@ -304,7 +304,7 @@ scDeadlock2 = do
   (,) <$> pure res <*> readMVar var
 
 -- | Subcomputation successfully completes.
-scSuccess :: Monad n => Conc n r (Either Failure ())
+scSuccess :: Monad n => ConcT r n (Either Failure ())
 scSuccess = do
   var <- newMVar ()
   subconcurrency $ do
@@ -313,7 +313,7 @@ scSuccess = do
     takeMVar out
 
 -- | Illegal usage
-scIllegal :: Monad n => Conc n r ()
+scIllegal :: Monad n => ConcT r n ()
 scIllegal = do
   var <- newEmptyMVar
   void . fork $ readMVar var
@@ -321,7 +321,7 @@ scIllegal = do
 
 -- | Test case from issue 71. This won't fail if the bug is
 -- reintroduced, it will just hang.
-scIssue71 :: Monad n => Conc n r ()
+scIssue71 :: Monad n => ConcT r n ()
 scIssue71 = do
   let ma ||| mb = do { j1 <- spawn ma; j2 <- spawn mb; takeMVar j1; takeMVar j2; pure () }
   s <- newEmptyMVar