+++ dejafu-tests/Cases/SingleThreaded.hs
@@ -12,7 +12,7 @@ import Test.HUnit (test)
 import Test.HUnit.DejaFu (testDejafu)
 
 import Control.Concurrent.Classy
-import Test.DejaFu.Conc (Conc, subconcurrency)
+import Test.DejaFu.Conc (ConcT, subconcurrency)
 
 import Utils
 
@@ -288,17 +288,17 @@ capsSet = do
 -- Subconcurrency
 
 -- | Subcomputation deadlocks.
-scDeadlock1 :: Monad n => Conc n r (Either Failure ())
+scDeadlock1 :: Monad n => ConcT r n (Either Failure ())
 scDeadlock1 = subconcurrency (newEmptyMVar >>= readMVar)
 
 -- | Subcomputation deadlocks, and action after it still happens.
-scDeadlock2 :: Monad n => Conc n r (Either Failure (), ())
+scDeadlock2 :: Monad n => ConcT r n (Either Failure (), ())
 scDeadlock2 = do
   var <- newMVar ()
   (,) <$> subconcurrency (putMVar var ()) <*> readMVar var
 
 -- | Subcomputation successfully completes.
-scSuccess :: Monad n => Conc n r (Either Failure ())
+scSuccess :: Monad n => ConcT r n (Either Failure ())
 scSuccess = do
   var <- newMVar ()
   subconcurrency (takeMVar var)